requirements.txt
colorama==0.3.9
cursor==1.1.0
decorator==4.1.2
dnspython==1.15.0
enum34==1.1.6
etaprogress==1.1.1
feedparser==5.2.1
halo==0.0.7
log-symbols==0.0.11
netaddr==0.7.19
phishtank==1.0.0
pyfiglet==0.7.5
python-dateutil==2.6.1
six==1.11.0
spinners==0.0.19
termcolor==1.1.0
urllib3==1.22
validators==0.12.0
